Mutation of the LowPassFilter to generic filter that can produce all types of filters
