+++ keywords.txt
@@ -17,12 +17,19 @@ setPin13High	KEYWORD2
 setPin13Low	KEYWORD2
 trailingZeros	KEYWORD2
 
+ResonantFilterNbits	KEYWORD1
 LowPassFilter	KEYWORD1
+HighPassFilter	KEYWORD1
+BandPassFilter	KEYWORD1
+NotchFilter	KEYWORD1
 setCutoffFreq	KEYWORD2
 setResonance	KEYWORD2
 next	KEYWORD2
 setCutoffFreqAndResonance	KEYWORD2
 LowPassFilter16	KEYWORD1
+HighPassFilter16	KEYWORD1
+BandPassFilter16	KEYWORD1
+NotchFilter16	KEYWORD1
 
 
 Oscil	KEYWORD1
